#178: Move DatePickerWithPopover to own file
